requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
httpx>=0.28
opencv-python>=4.10
numpy>=1.26
Pillow>=10.0
PyMuPDF>=1.24
CairoSVG>=2.7
